Finish `ugit commit` and `ugit log`
At the last time, we need to use exactly tree OID to do `read_tree`.
However, we need to store `commit` type just as what `git` does.

This commit does `ugit commit`. First, it is just a file stored in the
database with type `commit`. There are two parts of information:

1. PARENT COMMIT OID
2. USER CONTENT

Also, we need a `HEAD` file to store the current commit OID. And next we
use `ugit log` to iterate the commit ID just like a list.