Finish `ugit checkout`
`checkout` is used to set the working directory to commit OID. It just
use `read_tree` and set `HEAD` to OID. Now, we don't consider `ugit
branch` operation, so `HEAD` is not a symbolic file.

In the real `git`, `HEAD` is a symbolic file pointing to the current
branch. The actual commit OID is stored in `.git/refs/heads`.

However, the idea is simple but useful.