Finish `ugit tag` and refactor
Tag in git is just a symbolic reference to commit OID. In this commit,
we just use lightweight tag. It is simple.

However, we used `get_HEAD` and `set_HEAD` in the past. We should update
it into reference lookup. Because `HEAD` is also a reference file. So we
use `update_ref` and `get_ref` to refactor.

The user wants to use tag name to `ugit log` and so on. So it's better
to make it transparent by use `custom type` with `argparse`.