Finish `ugit branch` command
Branch name is just a reference which contains a commit OID. So in order
to create new branch, we need to create a file in `refs/heads/<name>`.
It is important to notice that `HEAD` file should be converted to
symbolic file.

So, in the past, we just used `update_ref` and `get_ref` to deal with
reference content. So we need to change to make it suitable for symbolic
file. It is a simple task.