Finish merge functionality
First we need to create `MERGE_HEAD` first which simply points to the
branch need to be merged. The simple idea is just compare `HEAD` tree
and `MERGE_HEAD` tree, fix the confict.

However, situation is more complicated. We'd rather find the two
branches' common ancestor. If the ancestor is just `HEAD` itself. We can
do fast-forward merge, just move `HEAD` to `MERGE_HEAD`.

If not, we use three-way merge by `diff3` command.